refactor(temporal): change global file tree organization
