+++ packages/temporal/mpe/workflows/mpe_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	shared_mpe "github.com/AdonisEnProvence/MusicRoom/shared/mpe"
+	shared_mpe "github.com/AdonisEnProvence/MusicRoom/mpe/shared"
 	"github.com/stretchr/testify/suite"
 	"go.temporal.io/sdk/workflow"
 )
