+++ packages/temporal/mtv/workflows/mtv.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/AdonisEnProvence/MusicRoom/activities"
-	shared_mtv "github.com/AdonisEnProvence/MusicRoom/shared/mtv"
+	shared_mtv "github.com/AdonisEnProvence/MusicRoom/mtv/shared"
 	"github.com/Devessier/brainy"
 
 	"github.com/mitchellh/mapstructure"
