+++ packages/temporal/mtv/workflows/mtv_events.go
@@ -1,7 +1,7 @@
 package mtv
 
 import (
-	shared_mtv "github.com/AdonisEnProvence/MusicRoom/shared/mtv"
+	shared_mtv "github.com/AdonisEnProvence/MusicRoom/mtv/shared"
 	"github.com/Devessier/brainy"
 )
